Fix Flat bedrock generator settings
