+++ build-data/paper.at
@@ -628,15 +628,7 @@ public net.minecraft.world.level.dimension.end.EnderDragonFight spawnNewGateway(
 public net.minecraft.world.level.entity.PersistentEntitySectionManager ensureChunkQueuedForLoad(J)V
 public net.minecraft.world.level.entity.PersistentEntitySectionManager permanentStorage
 public net.minecraft.world.level.gamerules.GameRules rules
-public net.minecraft.world.level.levelgen.SurfaceRules$Condition
-public net.minecraft.world.level.levelgen.SurfaceRules$Context
-public net.minecraft.world.level.levelgen.SurfaceRules$Context blockX
-public net.minecraft.world.level.levelgen.SurfaceRules$Context blockY
-public net.minecraft.world.level.levelgen.SurfaceRules$Context blockZ
-public net.minecraft.world.level.levelgen.SurfaceRules$Context context
-public net.minecraft.world.level.levelgen.SurfaceRules$Context randomState
-public net.minecraft.world.level.levelgen.SurfaceRules$LazyCondition
-public net.minecraft.world.level.levelgen.SurfaceRules$LazyYCondition
+public net.minecraft.world.level.levelgen.material.MaterialRuleContext context
 public net.minecraft.world.level.levelgen.structure.StructurePiece SHAPE_CHECK_BLOCKS
 public net.minecraft.world.level.levelgen.structure.placement.AbstractSpreadingStructurePlacement exclusionZone
 public net.minecraft.world.level.levelgen.structure.placement.AbstractSpreadingStructurePlacement frequency
