+++ paper-server/patches/sources/net/minecraft/server/Bootstrap.java.patch
@@ -8,15 +8,14 @@
              if (BuiltInRegistries.REGISTRY.keySet().isEmpty()) {
                  throw new IllegalStateException("Unable to load registries");
              }
-@@ -55,10 +_,84 @@
+@@ -55,10 +_,83 @@
              EntitySelectorOptions.bootStrap();
              DispenseItemBehavior.bootStrap();
              CauldronInteractions.bootStrap();
 -            BuiltInRegistries.bootStrap();
 +            // Paper start
-+            BuiltInRegistries.bootStrap(() -> {
-+                io.papermc.paper.world.worldgen.OptionallyFlatBedrockConditionSource.bootstrap(); // Paper - Flat bedrock generator settings
-+            });
++            // Paper - Flat bedrock generator settings
++            BuiltInRegistries.bootStrap(io.papermc.paper.world.worldgen.OptionallyFlatBedrockCondition::bootstrap);
 +            // Paper end
              CreativeModeTabs.validate();
              wrapStreams();
